Moved sodium 0.5.x to conflicts instead of breaks
You cannot break compatibility with one of the most used optimisation mods just because Flywheel, AN OPTIONAL COMPONENT, doesn't work right now. NOT EVEN A CRASH, just some stuff won't show up after reloading.

This is truly pathetic, and at least one issue about it has randomly been locked and had most of its comments marked as spam, at least one of which wrongly so. Other issues about this are marked as duplicates when there's a real problem you need to solve.
This has given me an AWFUL perception of TropheusJ, and the create fabric devteam as a whole.

I would be fine if it was in conflicts, because that's where it should be. BUT DON'T FULLY BREAK THE COMPAT FOR SUCH AN AVOIDABLE ISSUE!!!

TL;DR Moved sodium 0.5.x to conflicts because the reasoning for its inclusion in breaks can easily be sidestepped.